Refactored all 'reused test methods' into Helper classes
