+++ Marqeta.Core.Sdk.Tests/CardTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
 using Marqeta.Core.Abstractions;
-using Marqeta.Core.Sdk.Tests.Helpers;
+using Marqeta.Core.Sdk.Tests.Factories;
 using Xunit;
 
 // ReSharper disable IdentifierTypo
