+++ Marqeta.Core.Sdk.Tests/Factories/ClientFactory.cs
@@ -2,7 +2,7 @@
 
 // ReSharper disable IdentifierTypo
 
-namespace Marqeta.Core.Sdk.Tests.Helpers
+namespace Marqeta.Core.Sdk.Tests.Factories
 {
     public static class ClientFactory
     {
