+++ Marqeta.Core.Sdk.Tests/PingTests.cs
@@ -1,6 +1,6 @@
 ﻿using AutoFixture;
 using Marqeta.Core.Abstractions;
-using Marqeta.Core.Sdk.Tests.Helpers;
+using Marqeta.Core.Sdk.Tests.Factories;
 using Xunit;
 
 // ReSharper disable IdentifierTypo
