+++ Marqeta.Core.Sdk.Tests/ProgramFundingControlTests.cs
@@ -1,7 +1,4 @@
-﻿using System.Threading.Tasks;
-using AutoFixture;
-using Marqeta.Core.Abstractions;
-using Marqeta.Core.Sdk.Tests.Helpers;
+﻿using Marqeta.Core.Sdk.Tests.Helpers;
 using Xunit;
 
 // ReSharper disable IdentifierTypo
@@ -11,22 +8,9 @@ namespace Marqeta.Core.Sdk.Tests
     public class ProgramFundingControlTests : BaseTests
     {
         [Fact]
-        public async Task<Program_funding_source_response> FundingsourcesProgramPostAsync()
+        public async void FundingsourcesProgramPostAsync()
         {
-            // Get client / fixture
-            var client = ClientFactory.GetMarqetaClient();
-            var fixture = new Fixture();
-
-            // Create a program funding sources
-            var programFundingSourceRequest = fixture.Build<Program_funding_source_request>()
-                .Without(pfsr => pfsr.Active)
-                .Without(pfsr => pfsr.Token)
-                .Create();
-            var programFundingSourceResponse = await client.FundingsourcesProgramPostAsync(programFundingSourceRequest);
-            Assert.NotNull(programFundingSourceResponse);
-
-            // Return for use in other tests
-            return programFundingSourceResponse;
+            await ProgramFundingHelper.CreateProgramFundingSource();
         }
     }
 }
